Extracting FakeBeersRepository into a separate module to reuse.
