+++ .idea/gradle.xml
@@ -32,6 +32,7 @@
             <option value="$PROJECT_DIR$/beer_database" />
             <option value="$PROJECT_DIR$/beer_network" />
             <option value="$PROJECT_DIR$/beerdomain" />
+            <option value="$PROJECT_DIR$/beerdomain/test" />
             <option value="$PROJECT_DIR$/build-logic" />
             <option value="$PROJECT_DIR$/build-logic/convention" />
             <option value="$PROJECT_DIR$/buildSrc" />
