+++ beerdomain/src/test/java/com/simtop/beerdomain/domain/usecases/GetAllBeersUseCaseTest.kt
@@ -2,7 +2,7 @@ package com.simtop.beerdomain.domain.usecases
 
 import app.cash.turbine.test
 import com.simtop.beerdomain.domain.models.Beer
-import com.simtop.beerdomain.fakes.FakeBeersRepository
+import com.simtop.beerdomain.test.fakes.FakeBeersRepository
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
 import org.junit.jupiter.api.Test
