+++ beerdomain/src/test/java/com/simtop/beerdomain/domain/usecases/LoadNextPageUseCaseTest.kt
@@ -1,6 +1,6 @@
 package com.simtop.beerdomain.domain.usecases
 
-import com.simtop.beerdomain.fakes.FakeBeersRepository
+import com.simtop.beerdomain.test.fakes.FakeBeersRepository
 import com.simtop.core.core.PagingState
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
