+++ beerdomain/src/test/java/com/simtop/beerdomain/domain/usecases/ObservePagingStateUseCaseTest.kt
@@ -1,7 +1,7 @@
 package com.simtop.beerdomain.domain.usecases
 
 import app.cash.turbine.test
-import com.simtop.beerdomain.fakes.FakeBeersRepository
+import com.simtop.beerdomain.test.fakes.FakeBeersRepository
 import com.simtop.core.core.PagingState
 import kotlinx.coroutines.test.runTest
 import org.junit.jupiter.api.Assertions.assertEquals
